fiz só um teste porque estava dando muito erro e criei dois bancos de dados um que roda o app e outro que roda o teste.Explicação: os testes podem apagar dados do banco da aplicação e pode causar conflito se testar com a aplicação rodando.

+++ voce-aluga/src/main/java/voce/aluga/repository/ClienteRepository.java
@@ -3,7 +3,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import voce.aluga.model.Cliente;
 
+
+
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
+public interface ClienteRepository extends JpaRepository<Cliente, Long> {
     Cliente findByEmailAndSenha(String email, String senha);
+ 
 }
